requirements_v2.txt
langchain
langchain-community
langchain-ollama
langchain-core
oracledb
python-dotenv
streamlit
pypdf
python-docx
unstructured
docx2txt
